# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.AddParticipantRequest:
    properties:
      participant_id:
        type: string
      participant_type:
        example: agent
        type: string
    required:
    - participant_id
    - participant_type
    type: object
...
  handlers.CreateChatRequest:
    properties:
      chat_type:
        example: individual
        type: string
      contact_id:
        type: string
      creator_id:
        type: string
      subject:
        example: Team Discussion
        type: string
    required:
    - chat_type
    type: object
...
  handlers.UpdateSubjectRequest:
    properties:
      subject:
        example: New Team Discussion
        type: string
    required:
    - subject
    type: object
...
info:
  contact:
    email: support@ventros.com
    name: API Support
  description: API para gerenciamento de CRM com eventos e workflows
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Ventros CRM API
  version: "1.0"
paths:
...
  /api/v1/chats:
    get:
      consumes:
      - application/json
      description: List chats with optional filters (project_id, contact_id, status,
        chat_type)
      parameters:
      - description: Filter by project ID
        in: query
        name: project_id
        type: string
      - description: Filter by contact ID (participant)
        in: query
        name: contact_id
        type: string
      - description: Filter by status (active, archived, closed)
        in: query
        name: status
        type: string
      - description: Filter by chat type (individual, group, channel)
        in: query
        name: chat_type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of chats
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid parameters
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: List chats
      tags:
      - CRM - Chats
    post:
      consumes:
      - application/json
      description: Create a new chat (individual, group, or channel)
      parameters:
      - description: Project ID
        in: query
        name: project_id
        required: true
        type: string
      - description: Chat data
        in: body
        name: chat
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateChatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Chat created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new chat
      tags:
      - CRM - Chats
  /api/v1/chats/{id}:
    get:
      consumes:
      - application/json
      description: Get detailed information about a specific chat
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Chat details
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid chat ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Chat not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get chat by ID
      tags:
      - CRM - Chats
  /api/v1/chats/{id}/archive:
    post:
      consumes:
      - application/json
      description: Archive a chat (can be unarchived later)
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Chat archived successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid chat ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Chat not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Archive chat
      tags:
      - CRM - Chats
  /api/v1/chats/{id}/close:
    post:
      consumes:
      - application/json
      description: Permanently close a chat (cannot be reopened)
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Chat closed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid chat ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Chat not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Close chat
      tags:
      - CRM - Chats
  /api/v1/chats/{id}/participants:
    post:
      consumes:
      - application/json
      description: Add a contact or agent as a participant to a chat
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      - description: Participant data
        in: body
        name: participant
        required: true
        schema:
          $ref: '#/definitions/handlers.AddParticipantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Participant added successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Chat not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add participant to chat
      tags:
      - CRM - Chats
  /api/v1/chats/{id}/participants/{participant_id}:
    delete:
      consumes:
      - application/json
      description: Remove a participant from a chat
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      - description: Participant ID
        in: path
        name: participant_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Participant removed successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Chat or participant not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Remove participant from chat
      tags:
      - CRM - Chats
  /api/v1/chats/{id}/subject:
    patch:
      consumes:
      - application/json
      description: Update the subject/name of a group or channel chat
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      - description: New subject
        in: body
        name: subject
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateSubjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Subject updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Chat not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update chat subject
      tags:
      - CRM - Chats
  /api/v1/chats/{id}/unarchive:
    post:
      consumes:
      - application/json
      description: Unarchive a chat (reactivate it)
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Chat unarchived successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid chat ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Authentication required
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Chat not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Unarchive chat
      tags:
      - CRM - Chats
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: API Key authentication for service-to-service communication
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"